Homework 7 (#3)
