+++ Logs.js
@@ -1,3 +1,6 @@
+import {getRandom} from './getRandom.js';
+import {$chat} from './utils.js';
+
 
 export const logs = {
     start:
@@ -40,3 +43,45 @@ export const logs = {
     draw: "Ничья - это тоже победа!",
 };
 
+/**
+ * функция генирации лога, `<p>${text}</p>` - вариант использования шаблонноных строк, insertAdjacentHTML('afterbegin', el) - метод добовляет текст в зависимости от выбранной позиции
+ * @param {string} type
+ * @param {Object.<string, number>} player1
+ * @param {Object.<string, number>} player2
+ */
+export const generateLogs = (type, player1, player2) => {
+    const time = new Date().toLocaleTimeString();
+    let text = '';
+    const {hp:hp1, name:name1} = player1;
+    const {hp:hp2, name:name2} = player2;
+
+
+    switch (type) {
+        case 'hit':
+            text = `${time} - ${logs[type][getRandom(logs.hit.length)]
+                .replace('[playerKick]', name1)
+                .replace('[playerDefence]', name2)} ${hp2 - 100}[${
+                hp2}/100]`;
+            // console.log('###:hit', text);
+            break;
+        case 'defence':
+            text = `${time} - ${logs[type][getRandom(logs.hit.length)]
+                .replace('[playerKick]', name1)
+                .replace('[playerDefence]', name2)} ${hp1 - 100}[${
+                hp2}/100]`;
+            // console.log('###:def', text);
+            break;
+        case 'draw':
+            text = logs["draw"];
+            break;
+        case 'end':
+            text = `${time} - ${logs[type][getRandom(logs.end.length)]
+                .replace('[playerWins]', name1)
+                .replace('[playerLose]', name2)}`
+            break;
+        default:
+            text = logs[type].replace('time', time).replace('[player1]', name1).replace('[player2]', name2);
+    }
+    const el = `<p>${text}</p>`;
+    $chat.insertAdjacentHTML('afterbegin', el);
+}
