+++ utils.js
@@ -1,0 +1,8 @@
+
+export const $arenas = document.querySelector(".arenas");
+
+export const $submitButton = document.querySelector(".button");
+
+export const $formFight = document.querySelector(".control");
+
+export const $chat = document.querySelector('.chat');
